handle calculate rating category

+++ src/main/java/com/cghue/projecthousemaidwebapp/domain/FeedBack.java
@@ -23,7 +23,7 @@ public class FeedBack {
 
     private String description;
 
-    private Integer percent;
+    private Float percent;
 
     @ManyToOne
     private Order order;
